+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat: Autodetect more wizards (#370)
 - feat(apple): iOS wizard has support for cocoapods (#350)
 - feat(apple): Add Fastlane detector for iOS wizard (#356)
+- feat(sourcemaps): Add dedicated NextJS sourcemaps flow (#372)
 - fix(login): Avoid repeatedly printing loading message (#368)
 
 ## 3.7.1